refactor: 마이페이지 폴더 구조 변경
